pyproject.toml
[tool.poetry]
name = "dawith"
version = "0.1.1"
description = "다윗:작은이들의 외침"
authors = ["HansolLim <dev.hansollim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.1"
uvicorn = "^0.22.0"
gunicorn = "^20.1.0"
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"
jinja2-django-tags = "^0.5"
pymysql = "^1.1.0"
whitenoise = "^6.5.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
drf-spectacular = "^0.26.4"
[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
